requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
cloudinary
python-multipart
pymongo>=4.6
dnspython>=2.4